refactor(layout): rename shared/ to layers/shared_latent/ (T124.5.3)
Move the top-level shared/ package into layers/shared_latent/ and
rename the Go package from shared to shared_latent so the name reflects
its purpose (shared latent space layers). No external callers existed,
so this is a self-contained move plus package rename.

- git mv shared layers/shared_latent
- package shared -> package shared_latent in 3 files
- gofmt, go build ./..., go vet ./..., and package tests pass
- plan.md: T124.5.3 marked DONE 2026-04-27
